+++ packages/pangea-node-sdk/src/config.ts
@@ -1,6 +1,6 @@
 import { ConfigOptions, ConfigEnv } from "./types.js";
 
-export const version = "1.10.0";
+export const version = "2.1.0";
 
 class PangeaConfig {
   domain: string = "pangea.cloud";