+++ packages/react-mui-authn/src/views/AuthFlow/components/Signup/index.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import { useFormik } from "formik";
 import * as yup from "yup";
-import { Box, Stack, TextField, Typography } from "@mui/material";
+import { Stack, TextField, Typography } from "@mui/material";
 
 import { FlowStep } from "@pangeacyber/react-auth";
 
@@ -30,7 +30,9 @@ const SignupView: FC<ViewComponentProps> = ({
       .test(
         "password-requirements",
         "Password must meet requirements",
-        checkPassword
+        (value) => {
+          return checkPassword(value, data.passwordPolicy);
+        }
       ),
   });
 
@@ -54,7 +56,7 @@ const SignupView: FC<ViewComponentProps> = ({
   return (
     <Stack gap={2}>
       <Typography variant="h6">Signup</Typography>
-      <Typography variant="body2" sx={{ wordBreak: "break-word" }}>
+      <Typography variant="body2" mb={1} sx={{ wordBreak: "break-word" }}>
         Create an account with {data.email}
       </Typography>
       <form onSubmit={formik.handleSubmit}>
@@ -68,6 +70,7 @@ const SignupView: FC<ViewComponentProps> = ({
             onChange={formik.handleChange}
             error={formik.touched.firstName && Boolean(formik.errors.firstName)}
             helperText={formik.touched.firstName && formik.errors.firstName}
+            autoComplete="given-name"
           />
           <TextField
             fullWidth
@@ -78,6 +81,7 @@ const SignupView: FC<ViewComponentProps> = ({
             onChange={formik.handleChange}
             error={formik.touched.lastName && Boolean(formik.errors.lastName)}
             helperText={formik.touched.lastName && formik.errors.lastName}
+            autoComplete="family-name"
           />
           <PasswordField
             name="password"