+++ packages/react-mui-shared/src/components/PangeaDataGrid/components/Search/index.tsx
@@ -1,4 +1,4 @@
-import { FC, useRef, useState } from "react";
+import { FC, ReactNode, useRef, useState } from "react";
 import FilterAltOutlinedIcon from "@mui/icons-material/FilterAltOutlined";
 
 import {
@@ -28,6 +28,8 @@ interface SearchProps<FiltersObj> {
   Filters?: FilterFormProps<FiltersObj>;
   ColumnsPopoutProps?: ColumnsPopoutProps;
   EndFilterButton?: FC<FilterFormProps<FiltersObj>>;
+  EndBarComponent?: ReactNode;
+  StartBarComponent?: ReactNode;
 }
 
 const Search = <
@@ -41,6 +43,8 @@ const Search = <
   Filters,
   ColumnsPopoutProps,
   EndFilterButton,
+  EndBarComponent,
+  StartBarComponent,
 }: SearchProps<FiltersObj>): JSX.Element => {
   const [query_, setQuery_] = useInternalState(query, onChange);
   const searchRef = useRef<HTMLDivElement | null>(null);
@@ -49,7 +53,13 @@ const Search = <
 
   return (
     <Stack spacing={1}>
-      <Stack direction="row" spacing={1} sx={{ width: "100%" }}>
+      <Stack
+        className="PangeaDataGrid-SearchBar-root"
+        direction="row"
+        spacing={1}
+        sx={{ width: "100%" }}
+      >
+        {!!StartBarComponent && StartBarComponent}
         <ButtonGroup sx={{ flexGrow: 1 }}>
           <ConditionalAutocomplete
             ref={searchRef}
@@ -61,6 +71,7 @@ const Search = <
             onOpen={() => {
               setFilterMenuOpen(false);
             }}
+            size="small"
             InputProps={
               !!Filters
                 ? {
@@ -108,6 +119,7 @@ const Search = <
           Search
         </Button>
         {!!ColumnsPopoutProps && <ColumnsPopout {...ColumnsPopoutProps} />}
+        {!!EndBarComponent && EndBarComponent}
       </Stack>
       {!!Filters && Filters.showFilterChips && (
         <FiltersBar<FiltersObj> {...Filters} />