+++ packages/react-mui-shared/src/index.ts
@@ -9,11 +9,15 @@ export {
   ColumnsPopout,
   ColumnsPopoutProps,
   Visibility,
+  LinedPangeaDataGrid,
 } from "./components/PangeaDataGrid";
 export { default as JsonViewer } from "./components/JsonViewer";
 export { default as ConditionalAutocomplete } from "./components/ConditionalAutocomplete";
 export { default as PopoutCard } from "./components/PopoutCard";
 export { default as TimeRangeSelect } from "./components/TimeRangeSelect";
+
+/* Exported utils */
+export { getISO } from "./utils";
 
 /* Exported hooks */
 export { useInternalState } from "./utils/hooks";