+++ packages/react-mui-shared/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.23] - 2023-07-26
+
+### Changed
+
+- Upgraded mui x-date-pickers
+
+## [0.0.14] - 2023-03-28
+
 ## [0.0.16] - 2023-04-04
 
 ### Fixed